remove multipla parts selected on the tableview and confirmation

+++ src/main/java/com/partsinventory/controller/FetchPartsController.java
@@ -73,11 +73,15 @@ public class FetchPartsController {
             StackPane.setAlignment(searchGroup, Pos.CENTER);
             resultsStackPane.getChildren().add(tableViewRoot);
             deleteButton.setOnAction(e -> {
-                for(Part part : productTableView.getPartsListTableView().getSelectionModel().getSelectedItems()) {
-                    PartService.deletePart(part.getId());
-                    productTableView.getPartsListTableView().getSelectionModel().clearSelection();
+                if(PartService.handleDelete()){
+                    for(Part part : productTableView.getPartsListTableView().getSelectionModel().getSelectedItems()) {
+                        PartService.deletePart(part.getId());
+
+                    }
+                    //productTableView.getPartsListTableView().getSelectionModel().clearSelection();
                 }
-               // productTableView.getPartsListTableView().getSelectionModel().getSelectedItems().clear();
+
+
             });
         } catch (IOException e) {
             e.printStackTrace();
